terraform sqs development module + random admin_pass

+++ bin/teardown.sh
@@ -2,11 +2,14 @@
 
 functionUrl=$(aws ssm get-parameter --name /pwnctl/Api/BaseUrl | jq -r .Parameter.Value)
 
-adminPassword=$(aws secretsmanager get-secret-value --secret-id /aws/secret/pwnctl/admin_password | jq -r .SecretString)
+if [ ! -z $functionUrl ]
+then
+    adminPassword=$(aws secretsmanager get-secret-value --secret-id /aws/secret/pwnctl/admin_password | jq -r .SecretString)
 
-accessToken=$(curl -s -d '{"username":"admin","password":"'$adminPassword'"}' -H 'Content-Type: application/json' "${functionUrl}auth/grant" | jq -r .accessToken)
+    accessToken=$(curl -s -d '{"username":"admin","password":"'$adminPassword'"}' -H 'Content-Type: application/json' "${functionUrl}auth/grant" | jq -r .accessToken)
 
-# deletes all operations & EventBridge schedules
-curl -H "Authorization: Bearer $accessToken" -X DELETE "${functionUrl}ops"
+    # deletes all operations & EventBridge schedules
+    curl -H "Authorization: Bearer $accessToken" -X DELETE "${functionUrl}ops"
+fi
 
 terraform -chdir=infra destroy
